Improve top bar.
