+++ screens/Budget/BudgetScreen.tsx
@@ -1,7 +1,8 @@
 import { useQuery } from "@apollo/client";
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { Budget, BudgetCategory, GetBudgetsDocument, GetBudgetsQuery, GetMonthBreakdownDocument, GetMonthBreakdownQuery } from "../../components/generated";
-import { Button, SafeAreaView, StatusBar } from "react-native"
+import { Button, SafeAreaView, StatusBar, TouchableHighlight } from "react-native"
+import { AntDesign } from "@expo/vector-icons";
 
 import React from 'react';
 import { StyleSheet, View, Text, ScrollView } from 'react-native';
@@ -9,14 +10,14 @@ import { RootTabScreenProps } from "../../types";
 import { ChartDisplay } from "../../components/budget/ChartDisplay";
 import { BudgetList } from "../../components/budget/BudgetList";
 import { MissingBudget } from "../../components/budget/MissingBudget";
-import { TopBar } from "../../components/budget/TopBar";
 import { useRefresh } from "../../hooks/useRefresh";
 import { useAuth } from "../../hooks/useAuth";
+import { MONTHS_ORDER } from "../../constants/Months"
 
 
 export default function BudgetScreen({ navigation, route }: RootTabScreenProps<'Budget'>) {
     const passwordHash = useAuth();
-    const [month, setMonth] = useState("JULY")
+    const [month, setMonth] = useState("July")
     const [year, setYear] = useState(2022)
 
     const { data: budgetData, loading: budgetLoading, refetch: budgetRefetch } = useQuery<GetBudgetsQuery>(GetBudgetsDocument, {
@@ -30,7 +31,6 @@ export default function BudgetScreen({ navigation, route }: RootTabScreenProps<'
         }
     })
 
-
     useRefresh(() => {
         budgetRefetch()
         monthRefetch()
@@ -41,11 +41,67 @@ export default function BudgetScreen({ navigation, route }: RootTabScreenProps<'
     const plannedAmount = selectedBudget === undefined ? 0 :
         selectedBudget.budgetCategories?.reduce((previousValue, currentValue) => previousValue + currentValue.amount, 0)
 
+    const months = MONTHS_ORDER
+
+    const backAMonth = () => {
+        const curIndex = months.indexOf(month)
+        if (curIndex == 0) {
+            setMonth(months[months.length - 1])
+            setYear(year - 1)
+        } else {
+            setMonth(months[curIndex - 1])
+        }
+    }
+
+    const forwardAMonth = () => {
+        const curIndex = months.indexOf(month)
+        if (curIndex == months.length - 1) {
+            setMonth(months[0])
+            setYear(year + 1)
+        } else {
+            setMonth(months[curIndex + 1])
+        }
+    }
+
+    useEffect(() => {
+        navigation.setOptions({
+            headerTitle: `${month} ${year}`,
+            headerBackVisible: false,
+            headerLeft: (_) => {
+                return (
+                    <TouchableHighlight onPress={backAMonth} style={{
+                        alignItems: 'center',
+                        justifyContent: 'center',
+                        marginLeft: 10,
+                        width: 50,
+                        height: 50,
+                        borderRadius: 25,
+                    }} underlayColor="rgba(0,0,0,0.2)">
+                        <AntDesign name="left" size={32} color="black" />
+                    </TouchableHighlight>
+                );
+            },
+            headerRight: (_) => {
+                return (
+                    <TouchableHighlight onPress={forwardAMonth} style={{
+                        alignItems: 'center',
+                        justifyContent: 'center',
+                        marginRight: 10,
+                        width: 50,
+                        height: 50,
+                        borderRadius: 25,
+                    }} underlayColor="rgba(0,0,0,0.2)">
+                        <AntDesign name="right" size={32} color="black" />
+                    </TouchableHighlight>
+                );
+            },
+        });
+    }, [month]);
+
     return (
         <>
             <StatusBar />
             <SafeAreaView style={styles.container}>
-                <TopBar month={month} year={year} setMonth={setMonth} setYear={setYear} />
                 {(selectedBudget &&
                     <>
                         <ChartDisplay
